refactor(ui): extract token_format, key_reader, and banner_state modules
Fixes three layering issues identified in the post-PR re-review:

- token_format.py: extracts _CHARS_PER_TOKEN and format_token_count_short
  from streaming.py so runtime/state.py no longer imports from the ui
  streaming layer (inverted dependency).

- key_reader.py: consolidates duplicated raw terminal I/O (flush_stdin_unix,
  read_key_unix, read_key_windows) that was copied across choice_menu.py and
  feedback.py into a single shared module. Each caller adds its own thin
  action-mapping layer on top.

- banner_state.py: moves _SERVICE_DISPLAY_NAMES, _load_integration_health,
  _is_alert_listener_active, and _build_ambient_right_column out of banner.py
  (which already had banner_art and provider split off) so banner.py owns
  only rendering layout. banner.py: 505 → 382 lines.

Tests updated to monkeypatch banner_state module directly.
